added error control on the interpolation of slow states

+++ OMCompiler/SimulationRuntime/c/simulation/solver/gbode_ctrl.h
@@ -43,9 +43,9 @@
 extern "C" {
 #endif
 
-double CController(double* err_values, double* stepSize_values, double err_order);
-double IController(double* err_values, double* stepSize_values, double err_order);
-double PIController(double* err_values, double* stepSize_values, double err_order);
+double CController(double* err_values, double* stepSize_values, unsigned int err_order);
+double IController(double* err_values, double* stepSize_values, unsigned int err_order);
+double PIController(double* err_values, double* stepSize_values, unsigned int err_order);
 
 void gb_first_step(DATA* data, threadData_t* threadData, SOLVER_INFO* solverInfo);
